Replace ansii color codes with the formatter (#33)
* feat: :sparkles: Added more colors.

More colors have been added to the enum, as well as improved organization to make it easier to add more if needed in a structured way.

* feat(stylizer): extend color support in stylize function

Add support for standard ANSI colors and extended 256-color palette in the stylize function. Implement logic to distinguish between standard ANSI color codes and extended color codes for precise text styling in terminal output. This enhancement allows for a broader range of color customizations, improving the readability and visual appeal of terminal text.

* fix: Adapt code to new colors.

* refactor: :recycle: Replace ANSII color codes.

Replaced ANSII codes with our 'stylizer', and removed std::cout in favor of print and println.

* chore: :art: Formatting using clang-format.

* docs: :memo: Updated examples images for suite.

* refactor: Changed some outputs with newline.

* docs: :memo: Updated examples images for test suite.

* chore: Removed unneeded import.

---------

Co-authored-by: Alex Vergara <alex.vergara.dev@gmail.com>